Update NPC check to range-for, use pointer to NPC in DialogScreen
The check for an NPC near the player in GameScreen::talk() now uses a range for and passes a pointer to DialogScreen. Since DialogScreen no longer has to look up the NPC by index, cur_map (which holds the vector of NPCs) is once again a private member of GameScreen.