Require new button press to recognize action input
Previously, action commands were processed the same way as directional commands: a simple check if the key was down, and a delay in checking if we needed to give the player a chance to let the key up. However, if the player kept the key down too long, the action would be repeated. Now we require the last check to have shown the key up, so the player doesn't accidentally toggle an action repeatedly.